fix some bugs in test_reader_reset
